Move stack create handler to handler_stacks.go file
